cloud-run-execution/requirements.txt
gunicorn
flask
pyarrow
pandas
numpy
sklearn
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage